fix: uses parseFloat to handle borderWidth between 0 and 1
Sometimes depending on the browser zoom, the borderWidth can be between 0 and 1. Because of the parseInt, in these cases, the value was hidden, even though it was greater than 0.

Example, with a zoom of 150% and a borderWidth of 1px, the rendered borderWidth is 0.667px.

See https://bugzilla.mozilla.org/show_bug.cgi?id=1732769